submodules/agents/pyproject.toml
[tool.poetry]
name = "moragents"
version = "0.3.0"
description = "Server for moragents"
authors = ["Neo, Morpheus, and Trinity <neo@morpheus.com>"]
readme = "README.md"
include = ["src"]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
llama-cpp-python = "0.2.90"
werkzeug = "2.2.2"
web3 = "7.6.0"
scikit-learn = "1.5.1"
fastapi = ">=0.115.2,<1.0.0"
pymupdf = "1.22.5"
faiss-cpu = "1.9.0.post1"
feedparser = "*"
langchain-text-splitters = "0.3.0"
langchain-community = "0.3.10"
langchain-ollama = "0.2.1"
langchain-mcp-adapters = "0.0.11"
tweepy = "4.15.0"
uvicorn = "0.31.0"
python-dateutil = "2.9.0.post0"
python-multipart = "0.0.12"
beautifulsoup4 = "4.12.3"
selenium = "4.25.0"
cdp-sdk = "0.15.0"
aiofiles = "24.1.0"
pytz = "2024.2"
pyshorteners = "1.0.1"
webdriver-manager = "4.0.2"
psycopg2-binary = "2.9.10"
boto3 = "1.34.162"
together = "1.5.8"
crewai = {extras = ["tools"], version = "^0.118.0"}
pydantic = ">=2.5.3"
langchain-apify = "^0.1.2"

[tool.poetry.group.dev.dependencies]
pycodestyle = "^2.10.0"
pylint = ">=3.3.4"
autopep8 = "^2.0.2"
pytest = "^8.3.2"
pytest-mock = "^3.11.1"
pytest-asyncio = "^0.24.0"
mypy = "^1.4.1"
black = "^24.0.0"
alembic = "1.15.2"

[tool.poetry.scripts]
login = "login:run"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.12"
packages = ["src"]
namespace_packages = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = "sqlalchemy.*"
ignore_missing_imports = true

# This tells mypy to treat the src directory as a package root
[[tool.mypy.overrides]]
module = "src.*"
implicit_reexport = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
